* cleanups

* Fix adding /W4 in cmake function and cleanups.

Turns out functions make a new scope so setting /W4
wasn't sticking.  Turn it on and let the warnings flow
and then we can clean them up once we see them finally.

* compile without errors

* fix more warnings

* fix more warnings

* Don't use legacy rand/srand functions

Use better C++11 random functions, also simplify the logic
for creating the random password, allowing the posibility
for adding more characters into the default set easily if
desired.

* still fix more warnings

* Rename size_int to ssize in wwiv::stl

This is to match the std::ssize that will appear in C++20
support for compilers (it's in GCC 10, CLANG 9 so far, and
not yet added to MSVC).

* still fix more warnings

* Update regions.573

* Update to 810 regions file.

Update to 810 regions file.

* Don't use int for min/max values for input_number

That's too small to handle up to uint32_t which is used in WWIV.
This fixed #1205

* cleanup compiler warnings

+++ local_io/local_io.cpp
@@ -25,8 +25,8 @@ using namespace wwiv::strings;
 class DefaultCurAttrProvider final : public wwiv::local_io::curatr_provider {
 public:
   DefaultCurAttrProvider() = default;
-  [[nodiscard]] int curatr() const noexcept override { return a_; }
-  void curatr(int n) override { a_ = n; }
+  [[nodiscard]] uint8_t curatr() const noexcept override { return a_; }
+  void curatr(uint8_t n) override { a_ = n; }
 
 private:
   uint8_t a_{7};
@@ -41,8 +41,8 @@ LocalIO::~LocalIO() = default;
 void LocalIO::set_curatr_provider(wwiv::local_io::curatr_provider* p) { curatr_ = p; }
 wwiv::local_io::curatr_provider* LocalIO::curatr_provider() { return curatr_; }
 
-int LocalIO::curatr() const { return curatr_->curatr(); }
-void LocalIO::curatr(int c) { curatr_->curatr(c); }
+uint8_t LocalIO::curatr() const { return curatr_->curatr(); }
+void LocalIO::curatr(uint8_t c) { curatr_->curatr(c); }
 
 int LocalIO::EditLine(std::string& s, int len, AllowedKeys allowed_keys,
                     const std::string& allowed_set_chars) {
